refactor dispatcher modules to separate file
